tests/wpt/web-platform-tests/css/css-logical/parsing/inset-shorthand.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Logical Properties and Values: inset sets longhands</title>
<link rel="help" href="https://drafts.csswg.org/css-logical/#propdef-inset">
<meta name="assert" content="inset supports the full grammar '<'top'>{1,4}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
</head>
<body>
<script>
test_shorthand_value('inset', '1px 2px 3px 4px', {
  'top': '1px',
  'right': '2px',
  'bottom': '3px',
  'left': '4px'
});

test_shorthand_value('inset', '1px 2px 3px', {
  'top': '1px',
  'right': '2px',
  'bottom': '3px',
  'left': '2px'
});

test_shorthand_value('inset', '1px 2px', {
  'top': '1px',
  'right': '2px',
  'bottom': '1px',
  'left': '2px'
});

test_shorthand_value('inset', '1px', {
  'top': '1px',
  'right': '1px',
  'bottom': '1px',
  'left': '1px'
});
</script>
</body>
</html>
