LayoutTests/imported/w3c/web-platform-tests/css/css-logical/parsing/inset-shorthand-expected.txt


FAIL e.style['inset'] = "1px 2px 3px 4px" should set bottom assert_equals: bottom should be canonical expected "3px" but got ""
FAIL e.style['inset'] = "1px 2px 3px 4px" should set left assert_equals: left should be canonical expected "4px" but got ""
FAIL e.style['inset'] = "1px 2px 3px 4px" should set right assert_equals: right should be canonical expected "2px" but got ""
FAIL e.style['inset'] = "1px 2px 3px 4px" should set top assert_equals: top should be canonical expected "1px" but got ""
FAIL e.style['inset'] = "1px 2px 3px 4px" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['inset'] = "1px 2px 3px" should set bottom assert_equals: bottom should be canonical expected "3px" but got ""
FAIL e.style['inset'] = "1px 2px 3px" should set left assert_equals: left should be canonical expected "2px" but got ""
FAIL e.style['inset'] = "1px 2px 3px" should set right assert_equals: right should be canonical expected "2px" but got ""
FAIL e.style['inset'] = "1px 2px 3px" should set top assert_equals: top should be canonical expected "1px" but got ""
FAIL e.style['inset'] = "1px 2px 3px" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['inset'] = "1px 2px" should set bottom assert_equals: bottom should be canonical expected "1px" but got ""
FAIL e.style['inset'] = "1px 2px" should set left assert_equals: left should be canonical expected "2px" but got ""
FAIL e.style['inset'] = "1px 2px" should set right assert_equals: right should be canonical expected "2px" but got ""
FAIL e.style['inset'] = "1px 2px" should set top assert_equals: top should be canonical expected "1px" but got ""
FAIL e.style['inset'] = "1px 2px" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['inset'] = "1px" should set bottom assert_equals: bottom should be canonical expected "1px" but got ""
FAIL e.style['inset'] = "1px" should set left assert_equals: left should be canonical expected "1px" but got ""
FAIL e.style['inset'] = "1px" should set right assert_equals: right should be canonical expected "1px" but got ""
FAIL e.style['inset'] = "1px" should set top assert_equals: top should be canonical expected "1px" but got ""
FAIL e.style['inset'] = "1px" should not set unrelated longhands assert_true: expected true got false
